+++ sdk/core/azure-core-tracing-opencensus/azure/core/tracing/ext/opencensus_span/__init__.py
@@ -22,9 +22,20 @@ except ImportError:
     TYPE_CHECKING = False
 
 if TYPE_CHECKING:
-    from typing import Dict, Optional, Union, Callable
+    from typing import Dict, Optional, Union, Callable, Sequence
 
     from azure.core.pipeline.transport import HttpRequest, HttpResponse
+    AttributeValue = Union[
+        str,
+        bool,
+        int,
+        float,
+        Sequence[str],
+        Sequence[bool],
+        Sequence[int],
+        Sequence[float],
+    ]
+    Attributes = Optional[Dict[str, AttributeValue]]
 
 
 __version__ = VERSION
@@ -160,8 +171,8 @@ class OpenCensusSpan(HttpSpanMixin, object):
         return self.to_header()['traceparent']
 
     @classmethod
-    def link(cls, traceparent):
-        # type: (str) -> None
+    def link(cls, traceparent, attributes=None):
+        # type: (str, Attributes) -> None
         """
         Links the context to the current tracer.
 
@@ -170,11 +181,11 @@ class OpenCensusSpan(HttpSpanMixin, object):
         """
         cls.link_from_headers({
             'traceparent': traceparent
-        })
+        }, attributes)
 
     @classmethod
-    def link_from_headers(cls, headers):
-        # type: (Dict[str, str]) -> None
+    def link_from_headers(cls, headers, attributes=None):
+        # type: (Dict[str, str], Attributes) -> None
         """
         Given a dictionary, extracts the context and links the context to the current tracer.
 
@@ -183,7 +194,11 @@ class OpenCensusSpan(HttpSpanMixin, object):
         """
         ctx = trace_context_http_header_format.TraceContextPropagator().from_headers(headers)
         current_span = cls.get_current_span()
-        current_span.add_link(Link(ctx.trace_id, ctx.span_id))
+        current_span.add_link(Link(
+            trace_id=ctx.trace_id,
+            span_id=ctx.span_id,
+            attributes=attributes
+        ))
 
     @classmethod
     def get_current_span(cls):